diff: introduce Records
Introduce the `Records` struct which is used to wrap the records which
are to be compared with the `changed` slice which is used to keep track
if a record has been changed. Using a struct which is a wrapper around
slices instead of a slice of a (Record, bool) tuple gives the API a bit
more flexibility on the types which can be diffed. For example, the API
is much more efficient when diffing slices of raw u8's since we now
don't need to allocate a vector of references to each individual u8
(Vec<&u8>) and can instead just use the passed in &[u8]