merge: use correct range when outputting result
Fix a bug where the incorrect range was being used when outputting the
part of a merge where changes are present in both 'ours' and 'theirs',
e.g. the `MergeRange::Both` case. `MergeRange::Both` is a tuple of
(Range<'ours>, Range<'theirs>) and the range corresponding to 'ours' was
being used to index into the 'theirs' slice. Correct this by instead using the
'ours' range to index into the 'ours' slice.

Fixes: #9