apply: reduce the search space for finding a position to apply a hunk
This patch improves the worst case performance of finding a position to
apply a hunk by clamping the starting search position to the length of
the original image that is being patched. This bounds the search
algorithm to O(original image lines) instead of an unbounded number
based on the hunk header (which could be arbitrarily large and
incorrect).

The test case added in this patch would, before this patch, take >200ms
to run while it takes ~30us to run with this patch applied.

Fixes: #17